cryptkeeper: initial implementation
This is the initial implementation of cryptkeeper. This is a very large
PR as I have been working on this locally on my laptop without regularly
merging in code.

While I wouldn't recommend this approach, it made it easier for me to
implement and test locally. I'm writing this message for posterity and
to let possible future contributors understand the context.

This implementation has all the features that make cryptkeeper work,
including support for direnv and as a standalone. I may need to update
the flags around standalone support, but this has the base features,
unit tests and integration tests.